% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_diag_admin.R
\name{link_diag_admin}
\alias{link_diag_admin}
\title{Link diagnostic and demographic datasets using unique personal identifiers}
\usage{
link_diag_admin(
  data_diag,
  data_admin_inv = NULL,
  data_admin_var = NULL,
  id_col = "id",
  date_col = "year",
  log_path = NULL
)
}
\arguments{
\item{data_diag}{A data frame containing pre-processed and pre-validated diagnostic data.}

\item{data_admin_inv}{A data frame containing validated time-invariant administrative (sociodemographic) data.}

\item{data_admin_var}{A data frame containing validated time-variant administrative (sociodemographic) data.}

\item{id_col}{A character string. Name of ID (unique personal identifier) column in all of the provided datasets. Default is "id".}

\item{date_col}{A character string. Name  of the date column in time-variant data and diagnostic data.}

\item{log_path}{A character string. Path to the log file to append function logs. Default is \code{NULL}.
\itemize{
\item If \code{NULL}, a new directory \verb{/log} and file is created in the current working directory.
}}
}
\value{
Linked dataset including relevant diagnostic and administrative (sociodemographic) variables.
}
\description{
Link diagnostic and demographic datasets using unique personal identifiers
}
\examples{
# Link diagnostic and time invariant datasets
log_file <- tempfile()
cat("Example log file", file = log_file)

linked_diag_inv <- link_diag_admin(data_diag = diag_df,
                                  data_admin_inv = invar_df,
                                  id_col = "id",
                                  log_path = log_file)

# Link diagnostic and time variant datasets
names(var_df)[names(var_df) == 'year_varying'] <- 'year'
names(diag_df)[names(diag_df) == 'diag_year'] <- 'year'

linked_diag_var <- link_diag_admin(data_diag = diag_df,
                                  data_admin_var = var_df,
                                  id_col = "id",
                                  date_col = "year",
                                  log_path = log_file)

# Link diagnostic, time invariant and variant datasets
linked_diag_inv_var <- link_diag_admin(data_diag = diag_df,
                                      data_admin_var = var_df,
                                      data_admin_inv = invar_df,
                                      id_col = "id",
                                      date_col = "year",
                                      log_path = log_file)

}
