% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map.R
\name{plot_map}
\alias{plot_map}
\title{Plot spatial data}
\usage{
plot_map(
  data,
  level = c("kommune", "fylke"),
  rate_col,
  region_number_col,
  legend_title = "",
  map_title = "",
  percent = FALSE,
  auto_transform = FALSE
)
}
\arguments{
\item{data}{A data frame with the prevalence/incidence rates and auxiliary information.}

\item{level}{A character string. Type geographical level, can be "kommune" or "fylke".
\itemize{
\item Kommune 2025 geojson file retrieved from https://kartkatalog.geonorge.no/metadata/administrative-enheter-kommuner/041f1e6e-bdbc-4091-b48f-8a5990f3cc5b
\item Fylke 2024 geojson file retrieved from https://kartkatalog.geonorge.no/metadata/administrative-enheter-fylker/6093c8a8-fa80-11e6-bc64-92361f002671
}}

\item{rate_col}{A character string.  Name of the column in \code{data} containing the information (e.g. rates) to map.}

\item{region_number_col}{A character string. Name of the column in \code{data} containing fylke or kommune identification numbers in 2024 standard (harmonized series). For more information please consult the vignette("other_useful_fun")}

\item{legend_title}{A character string. Title for the legend box.}

\item{map_title}{Character string. Title of the plot.}

\item{percent}{Logical. Do the numbers in rate_col represent proportions or percentages? If \code{TRUE} formats scale as percentage. Default is \code{FALSE}.}

\item{auto_transform}{Logical. Sometimes if your data is skewed, the color palette and scale in the map might not look right. If \code{TRUE}, applies log10 transformation. Default is \code{FALSE}.}
}
\value{
Choropleth map
}
\description{
Plot spatial data
}
\examples{
pop_kommuner_2024 <- get_population_ssb(
regions = "kommuner",
years = 2024,
ages = c(0:18),
aggregate_age = TRUE,
by_sex = FALSE)

pop_kommuner_2024 <- harmonize_municipality_codes(
data = pop_kommuner_2024,
municipality_col = "region_code") |>
dplyr::filter(age == "Total")

plot_map(pop_kommuner_2024,
level = "kommune",
rate_col = "population",
map_title = "Population children in Norway 2024",
legend_title = "Population",
region_number_col = "harmonized_code",
auto_transform = TRUE)
}
