% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regtools-package.R
\docType{package}
\name{regtools-package}
\alias{regtools}
\alias{regtools-package}
\title{regtools: regtools: facilitating manipulation, analysis and visualization of data from Norwegian health and population registers.}
\description{
R package to facilitate the manipulation, analysis and visualization of data from Norwegian health and population registers.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/amslala/regtools}
  \item \url{https://amslala.github.io/regtools/}
  \item Report bugs at \url{https://github.com/amslala/regtools/issues}
}

}
\keyword{internal}
